\name{permPath-package}
\alias{permPath-package}
\alias{permPath}
\docType{package}
\title{
What the package does (short line)
 title
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab permPath\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-10-05\cr
License: \tab What license is it under?\cr
}
 license
}
\author{
Author: Ivo D. Shterev <i.shterev@duke.edu>

Maintainer: Ivo D. Shterev <i.shterev@duke.edu>
}
\references{
 literature
}

\keyword{ package }
\seealso{
 seealso
}
\examples{
 example
}
