\name{permPath-package}
\alias{permPath-package}
\alias{permPath}
\docType{package}
\title{
Permutation Based Gene Expression Pathway Analysis.
}
\description{
Can be used to carry out permutation based gene expression pathway analysis. This work was supported by a National Institute of Allergy and Infectious Disease/National Institutes of Health contract (No. HHSN272200900059C).
}
\details{
\tabular{ll}{
Package: \tab permPath\cr
Type: \tab Package\cr
Version: \tab 0.6\cr
Date: \tab 2016-05-13\cr
License: \tab GPL-3\cr
}
}
\author{
  I. D. Shterev, K. Owzar and G. D. Sempowski

Maintainer: I. D. Shterev <i.shterev@duke.edu>
}
\references{
   B. Efron, R. Tibshirani (2007) 
   On Testing the Significance of Sets of Genes. \emph{The Annals 
   of Applied Statistics}. \bold{Vol. 1}, No 1, 107--129.

   A. Subramanian, P. Tamayo, V. K. Mootha, S. Mukherjee, 
   B. L. Ebert, M. A. Gillette, A. Paulovich, S. L. Pomeroy, 
   T. R. Golub, E. S. Lander and J. P. Mesirov (2005), 
   Gene Set Enrichment Analysis: A knowledge-based Approach 
   for Interpreting Genome-Wide Expression Profiles. 
   \emph{Proc. Natl. Acad. Sci. USA}. \bold{Vol. 102}, 
   No 43, 15545--15550.
}
